Feat : 토스트 기능 추가 / Fix : 모달 닫힌 상태에서 스크롤 안됨 해결

+++ components/modal/DeleteFolderModal.tsx
@@ -2,6 +2,8 @@ import useModalStore from "@/store/useModalStore";
 import SubmitButton from "../SubMitButton";
 import ModalContainer from "./modalComponents/ModalContainer";
 import { deleteFolder } from "@/lib/api/folder";
+import toast from "react-hot-toast";
+import toastMessages from "@/lib/toastMessage";
 
 const DeleteFolderModal = ({
   folderName,
@@ -11,13 +13,20 @@ const DeleteFolderModal = ({
   folderId: number;
 }) => {
   const { closeModal } = useModalStore();
+  let linkCount: number;
   const handleSubmit = async () => {
-    // 폴더 내에 링크 개수 0 일때만 삭제 가능 -> 링크 1개 이상이면 error toast 띄우겠습니다.
-    try {
-      await deleteFolder(folderId);
-    } catch (error) {
-      console.log(error, "DeleteFolderModal 폴더 삭제 에러");
-    } finally {
+    // 폴더 내에 링크 개수 0 일때만 폴더 삭제 가능 -> 링크 1개 이상이면 error toast 띄우고 있음 or 전체 링크 삭제 후 폴더 삭제
+    if (linkCount === 0) {
+      try {
+        await deleteFolder(folderId);
+        toast.success(toastMessages.success.deleteFolder);
+      } catch (error) {
+        toast.error(toastMessages.error.deleteFolder);
+      } finally {
+        closeModal();
+      }
+    } else {
+      toast.error(toastMessages.error.deleteNonEmptyFolder);
       closeModal();
     }
   };
